Finish work on app layout
