+++ packages/frontend/src/components/app/app.jsx
@@ -6,7 +6,7 @@ import LandingPage from '../landingPage';
 import NotFoundPage from '../notFoundPage';
 import ConnectedLogin from '../login';
 import Signup from '../signup';
-import Dashboard from '../dashboard';
+import ConnectedDashboard from '../dashboard';
 import PrivateRoute from '../privateRoute';
 
 
@@ -15,7 +15,7 @@ const App = () => (
     <Route exact path={routes.home} component={LandingPage} />
     <Route path={routes.login} component={ConnectedLogin} />
     <Route path={routes.signup} component={Signup} />
-    <PrivateRoute component={Dashboard} path={routes.dashboard} />
+    <PrivateRoute path={routes.dashboard} component={ConnectedDashboard} />
     <Route component={NotFoundPage} />
   </Switch>
 );
