+++ packages/frontend/src/components/dashboard/dashboard.scss
@@ -4,21 +4,6 @@
 
   .dashboard-content {
     padding: 10px;
-    min-height: 100vh;
-
-    .user-menu {
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      margin-left: auto;
-      padding: 30px 0;
-      width: 200px;
-
-      p.user_name {
-        color: inherit;
-        margin-right: 10px;
-      }
-    }
 
     .content-area {
       max-width: 90%;
@@ -158,131 +143,6 @@
         .summary-description__sub-text {
           font-size: 18px;
         }
-      }
-    }
-  }
-
-  .sidebar-toggler {
-    position: fixed;
-    height: 51px;
-    width: 51px;
-    outline: none;
-    left: 5%;
-    top: 30px;
-    z-index: 1;
-  }
-
-  .button.sidebar-trigger {
-    background-color: #2d3039;
-    border: 1px solid #ffffff44;
-    border-radius: 50px;
-    cursor: pointer;
-    height: 50px;
-    width: 50px;
-    outline: none;
-    -webkit-box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.14), 0 2px 1px -1px rgba(0, 0, 0, 0.12),
-      0 1px 3px 0 rgba(0, 0, 0, 0.2);
-    box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.14), 0 2px 1px -1px rgba(0, 0, 0, 0.12), 0 1px 3px 0 rgba(0, 0, 0, 0.2);
-
-    .sidebar-trigger__icon {
-      /* icon created in CSS */
-      position: absolute;
-      left: 50%;
-      top: 50%;
-      bottom: auto;
-      right: auto;
-      -webkit-transform: translateX(-50%) translateY(-50%);
-      -moz-transform: translateX(-50%) translateY(-50%);
-      -ms-transform: translateX(-50%) translateY(-50%);
-      -o-transform: translateX(-50%) translateY(-50%);
-      transform: translateX(-50%) translateY(-50%);
-      width: 22px;
-      height: 2px;
-      background-color: #ffffff;
-    }
-
-    .sidebar-trigger__icon::before,
-    .sidebar-trigger__icon::after {
-      /* upper and lower lines of the menu icon */
-      content: "";
-      position: absolute;
-      top: 0;
-      right: 0;
-      width: 100%;
-      height: 100%;
-      background-color: inherit;
-      /* Force Hardware Acceleration in WebKit */
-      -webkit-transform: translateZ(0);
-      -moz-transform: translateZ(0);
-      -ms-transform: translateZ(0);
-      -o-transform: translateZ(0);
-      transform: translateZ(0);
-      -webkit-backface-visibility: hidden;
-      backface-visibility: hidden;
-      -webkit-transition: -webkit-transform 0.5s, width 0.5s, top 0.3s;
-      -moz-transition: -moz-transform 0.5s, width 0.5s, top 0.3s;
-      transition: transform 0.5s, width 0.5s, top 0.3s;
-    }
-
-    .sidebar-trigger__icon::before {
-      -webkit-transform-origin: right top;
-      -moz-transform-origin: right top;
-      -ms-transform-origin: right top;
-      -o-transform-origin: right top;
-      transform-origin: right top;
-      -webkit-transform: translateY(-6px);
-      -moz-transform: translateY(-6px);
-      -ms-transform: translateY(-6px);
-      -o-transform: translateY(-6px);
-      transform: translateY(-6px);
-    }
-
-    .sidebar-trigger__icon::after {
-      -webkit-transform-origin: right bottom;
-      -moz-transform-origin: right bottom;
-      -ms-transform-origin: right bottom;
-      -o-transform-origin: right bottom;
-      transform-origin: right bottom;
-      -webkit-transform: translateY(6px);
-      -moz-transform: translateY(6px);
-      -ms-transform: translateY(6px);
-      -o-transform: translateY(6px);
-      transform: translateY(6px);
-    }
-  }
-
-  .sidebar-toggler.sidebar-open {
-    .button.sidebar-trigger {
-      /* rotate trigger when navigation becomes visible */
-      -webkit-transform: rotate(180deg);
-      -moz-transform: rotate(180deg);
-      -ms-transform: rotate(180deg);
-      -o-transform: rotate(180deg);
-      transform: rotate(180deg);
-
-      .sidebar-trigger__icon::after,
-      .sidebar-trigger__icon::before {
-        /* animate arrow --> from hamburger to arrow */
-        width: 50%;
-        -webkit-transition: -webkit-transform 0.5s, width 0.5s;
-        -moz-transition: -moz-transform 0.5s, width 0.5s;
-        transition: transform 0.5s, width 0.5s;
-      }
-
-      .sidebar-trigger__icon::before {
-        -webkit-transform: rotate(45deg);
-        -moz-transform: rotate(45deg);
-        -ms-transform: rotate(45deg);
-        -o-transform: rotate(45deg);
-        transform: rotate(45deg);
-      }
-
-      .sidebar-trigger__icon::after {
-        -webkit-transform: rotate(-45deg);
-        -moz-transform: rotate(-45deg);
-        -ms-transform: rotate(-45deg);
-        -o-transform: rotate(-45deg);
-        transform: rotate(-45deg);
       }
     }
   }
