+++ packages/frontend/src/components/dashboard/index.js
@@ -1,1 +1,2 @@
 export { default } from './dashboard';
+export { Dashboard } from './dashboard';
