+++ packages/frontend/src/components/error/error.scss
@@ -1,28 +1,46 @@
 .error-page-wrapper {
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+  position: absolute;
+  top: 40%;
+  width: 100%;
 
-  .content-wrapper {
-    max-width: 500px;
-    text-align: center;
+  .error-page_inner {
+    
+    .content-wrapper {
+      text-align: center;
+      
+      .error-message {
+        font-size: 1.5rem;
+        text-transform: uppercase;
+      }
+      
+      .error-description {
+        color: #b1b1b1;
+        margin: auto;
+        max-width: 500px;
+        margin: 1rem auto;
+      }
 
-    .error-message {
-      font-size: 1.5rem;
-      text-transform: uppercase;
-      margin-bottom: 1rem;
+      .try-again {
+        text-transform: uppercase;
+        color: #d70508;
+        font-weight: 600;
+      }
     }
+  }
+}
 
-    .error-description {
-      color: #b1b1b1;
-      margin-bottom: 20px;
+@media screen and (max-width: 768px) {
+  .error-page-wrapper {
+    .content-wrapper {
+      padding: 0 20px;
     }
+  }
+}
 
-    .try-again {
-      text-transform: uppercase;
-      color: #000000;
-      font-weight: 600;
+@media screen and (min-width: 769px) {
+  .error-page-wrapper {
+    .content-wrapper {
+      margin-left: 250px;
     }
   }
 }
