+++ packages/frontend/src/components/loader/loader.jsx
@@ -2,8 +2,10 @@ import React from 'react';
 import './loader.scss';
 
 const Loader = () => (
-  <div className="wrapper">
-    <div className="loader" />
+  <div className="overlay">
+    <div className="overlay_inner">
+      <div className="loader" />
+    </div>
   </div>
 );
