+++ packages/frontend/src/components/loader/loader.scss
@@ -1,15 +1,22 @@
-.wrapper {
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.overlay {
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  position: absolute;
 
-  .loader {
-    width: 40px;
-    height: 40px;
-    background-color: #d70508;
-    border-radius: 100%;
-    animation: pulse 1.2s infinite cubic-bezier(0.455, 0.03, 0.515, 0.955);
+  .overlay_inner {
+    position: absolute;
+    left: 40%;
+    top: 40%;
+
+    .loader {
+      width: 50px;
+      height: 50px;
+      background-color: #d70508;
+      border-radius: 100%;
+      animation: pulse 1.2s infinite cubic-bezier(0.455, 0.03, 0.515, 0.955);
+    }
   }
 
   @keyframes pulse {
@@ -20,5 +27,11 @@
       transform: scale(1);
       opacity: 0;
     }
+  }
+}
+
+@media screen and (min-width: 769px) {
+  .overlay {
+    margin-left: 250px;
   }
 }
