+++ packages/frontend/src/components/mainLayout/index.js
@@ -1,0 +1,1 @@
+export { default } from './mainLayout';
