+++ packages/frontend/src/components/privateRoute/privateRoute.test.js
@@ -1,8 +1,12 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { MemoryRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import configureStore from 'redux-mock-store';
 
 import PrivateRoute from './privateRoute';
+
+const mockStore = configureStore([]);
 
 describe('56730186', () => {
   it('should render component if user has been authenticated', () => {
@@ -14,9 +18,16 @@ describe('56730186', () => {
     localStorage.setItem('user', JSON.stringify(userObj));
     const AComponent = () => <div>AComponent</div>;
     const props = { component: AComponent, path: '/dashboard' };
+    const store = mockStore({
+      global: {
+        currentUser: { id: 1 },
+      },
+    });
     const wrapper = mount(
       <MemoryRouter initialEntries={[props.path]}>
-        <PrivateRoute {...props} />
+        <Provider store={store}>
+          <PrivateRoute {...props} />
+        </Provider>
       </MemoryRouter>,
     );
     expect(wrapper.exists(AComponent)).toBe(true);
